BUGFIX: Don't convert route options to strings when merging subroutes
Adjusts the `RouteConfigurationProcessor` so that it doesn't implicitly
convert route configuration options to strings when merging routes.

Background:

This fixes a regression introduced with #2145 that lead to the Neos
route

```yaml
-
  name: 'Homepage'
  uriPattern: '{node}'
  routeParts:
    'node':
      options:
        onlyMatchSiteNodes: true
```

to be converted such that the `onlyMatchSiteNodes` option was casted
to a string of `"1"` leading to the homepage route resolving all node links.

Resolves: #2142