BUGFIX: Fix default order of middleware components
Adjusts the order of the Middleware components so that
the `SessionMiddleware` is executed before the `RoutingMiddleware`.

Otherwise session based authentication won't work until the
routing middleware was executed.

This also removes most of the explicit `position` configurations
in order to avoid too much interdependency.

If a 3rd party middleware needs to be executed before/after another
one, it can still use `position: before/after <name>`  of course.
Depending on the order of _multiple_ other components is considered
bad practice. But if that's really required one could still add a
`position` setting to the existing middleware configuration.

Related: #2019