BUGFIX: Fix console tests on windows machines and GitHub Actions
For unkown reason the "user" input is not be included in the
resulting output stream within our test scenarios on windows machines
and GitHub Actions.

In the end we removed the plain output comparision altogether
since the resulting output comes from
``Symfony\Component\Console\QuestionChoiceQuestion``
and we do not have to test internals of ``symfony/console`` anyway.