BUGFIX: Prevent recursive instanciation of EntityManager
With the changes in #2423 the PersistenceManager was registered as a Doctrine EventListener.
Hence, when building the EntityManager, the PersistenceManager got instanciated with a completely new non-lazy EntityManager. This caused errors when trying to persist entities, as they were not known to that instance of the EntityManager.
This change extracts the EventListener behaviour from the PersistenceManager.