BUGFIX: Correctly set up TestingProvider for "testable security"
As of https://github.com/neos/flow-development-collection/pull/1213
the authentication providers are created using a factory method, this
breaks the assumption the `TestingProvider` is a singleton.

The result is that e.g. ´authenticateRoles([])` no longer works as
expected in functional tests.

This fixes the issue by fetching the provider from the factory instead
of the object manager.

Fixes #2386