TASK: Add configuration file for RTD
This fixes an error caused by recent docutils and old sphinx being
incompatible.

The location of the configuration file in the root of the development
collection allows to (finally!) build the documentation directly from
the collection repository instead of the neos/flow split repository.

That finally fixes the "edit on GH" links, too.