BUGFIX: Reimplement request method override via middleware
The documented feature to override the request method with the POST argument `__method` or one of the HTTP headers X-Http-Method-Override or X-Http-Method was accidentally removed in the transition towards PSR-7. This adds it back in by adding a middleware that does the same logic that was removed with the previous Http\Request class.