BUGFIX: Avoid errors when `./flow help` is shown for an action without doc-block
The `./flow help` command currently throws errors when rendering details for command that has no doc-block.
Since modern php and flow need those doc-blocks less and less this should be supported aswell.

This change will return the argument name as description when no @param annotation is found for a parameter.
This is not as helpful as a description but shows the parameter and avoids an unnecessary error.