BUGFIX: Create serialization code for transient properties
Due to a recent optimization, Flow was not generating `__sleep()` methods for classes which are not either entities or were configured with a session scope. This led to errors in classes which were using the `@Transient` annotation to exclude certain properties from serialization. Therefore, Flow now also generates proxy classes with `__sleep()` methods if the original class contains such annotations.

Resolves: #3062